requirements.txt
google-cloud-firestore
faker